+++ .gitignore
@@ -5,3 +5,4 @@
 target
 .metals
 spark-warehouse
+.bsp
