--- ProjectExample/src/main/scala/com/leobenkel/zparkioProjectExample/Main.scala
@@ -1,3 +1,0 @@
-package com.leobenkel.zparkioProjectExample
-
-object Main extends Application {}
