--- ProjectExample_MoreComplex/src/main/scala/com/leobenkel/zparkioProfileExampleMoreComplex/Items/Post.scala
@@ -1,8 +1,0 @@
-package com.leobenkel.zparkioProfileExampleMoreComplex.Items
-
-case class Post(
-  postId:   Int,
-  authorId: Int,
-  title:    String,
-  content:  String
-)
