--- ProjectExample_MoreComplex/src/main/scala/com/leobenkel/zparkioProfileExampleMoreComplex/Items/User.scala
@@ -1,8 +1,0 @@
-package com.leobenkel.zparkioProfileExampleMoreComplex.Items
-
-case class User(
-  userId: Int,
-  name:   String,
-  age:    Int,
-  active: Boolean
-)
