--- ProjectExample_MoreComplex/src/main/scala/com/leobenkel/zparkioProfileExampleMoreComplex/Main.scala
@@ -1,3 +1,0 @@
-package com.leobenkel.zparkioProfileExampleMoreComplex
-
-object Main extends Application {}
