+++ examples/Example1_mini/src/main/scala/com/leobenkel/example1/Log.scala
@@ -1,0 +1,13 @@
+package com.leobenkel.example1
+
+import com.leobenkel.zparkio.Services._
+import zio.Task
+import zio.console.Console
+
+case class Log(console: Console.Service) extends Logger.Service {
+  override def info(txt: => String): Task[Unit] = console.putStrLn(s"INFO: $txt")
+
+  override def error(txt: => String): Task[Unit] = console.putStrLn(s"ERROR: $txt")
+
+  override def debug(txt: => String): Task[Unit] = console.putStrLn(s"DEBUG: $txt")
+}
