+++ examples/Example1_mini/src/main/scala/com/leobenkel/example1/Main.scala
@@ -1,0 +1,3 @@
+package com.leobenkel.example1
+
+object Main extends Application {}
