+++ examples/Example1_mini/src/test/resources/log4j.properties
