+++ examples/Example2_small/src/main/scala/com/leobenkel/example2/Items/Post.scala
@@ -1,0 +1,8 @@
+package com.leobenkel.example2.Items
+
+case class Post(
+  postId:   Int,
+  authorId: Int,
+  title:    String,
+  content:  String
+)
