+++ examples/Example2_small/src/main/scala/com/leobenkel/example2/Items/User.scala
@@ -1,0 +1,8 @@
+package com.leobenkel.example2.Items
+
+case class User(
+  userId: Int,
+  name:   String,
+  age:    Int,
+  active: Boolean
+)
