+++ examples/Example2_small/src/main/scala/com/leobenkel/example2/Main.scala
@@ -1,0 +1,3 @@
+package com.leobenkel.example2
+
+object Main extends Application {}
