+++ examples/Example2_small/src/main/scala/com/leobenkel/example2/Services/SparkBuilder.scala
@@ -1,0 +1,17 @@
+package com.leobenkel.example2.Services
+
+import com.leobenkel.zparkio.Services.SparkModule
+import com.leobenkel.example2.Arguments
+import org.apache.spark.sql.SparkSession
+
+object SparkBuilder extends SparkModule.Factory[Arguments] {
+  lazy final override protected val appName: String = "Zparkio_test"
+
+  override protected def updateConfig(
+    sparkBuilder: SparkSession.Builder,
+    arguments:    Arguments
+  ): SparkSession.Builder = {
+    sparkBuilder
+      .config("spark.foo.bar", arguments.sparkConfig())
+  }
+}
