+++ examples/Example2_small/src/test/resources/log4j.properties
