+++ scripts/spark-cross-clean.sh
