+++ scripts/spark-cross-publish.sh
